Decouple libConfiguration from libCommon
The dependency of libConfiguration to libCommon is not actually required
as the only usage of libCommon is done by helper tools, not
by libConfiguration itself. This takes that into account and
decouples the two libraries. This is required by the fact that
currently DataSampling has a dependency on libConfiguration but not
libCommon and this simplifies therefore integration.